pyproject.toml
[project]
name = "langchain-ai-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "black>=26.1.0",
    "chromadb>=1.5.5",
    "isort>=7.0.0",
    "langchain>=1.2.10",
    "langchain-chroma>=1.1.0",
    "langchain-classic>=1.0.2",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.19",
    "langchain-openai>=1.1.10",
    "langchain-tavily>=0.2.17",
    "langchain-text-splitters>=1.1.1",
    "langchainhub>=0.1.21",
    "langgraph>=1.0.8",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
    "tavily-python>=0.7.23",
]
